RokTracker/requirements_win64.txt
# RokTracker (Windows x64) - runtime deps
# Focus: scanners + Title Bot + debug tools

numpy>=1.26,<3
opencv-python>=4.8,<5
Pillow>=10,<12
pytesseract>=0.3.10,<1
requests>=2.31,<3

# Scanners output
pandas>=2.1,<3
openpyxl>=3.1,<4
XlsxWriter>=3.2,<4

# UI helpers
customtkinter>=5.2,<6
pathvalidate>=3.2,<4

# Optional (may require MSVC Build Tools on Windows):
# tesserocr
